pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "noslouchbench"
version = "0.1.0"
description = "Reproducible posture detection benchmarking framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mediapipe>=0.10.0",
  "opencv-python>=4.8.0",
  "numpy>=1.24.0",
  "PyYAML>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
